Xtreme Ini Control for Debug Purposes
Primarily meant for debugging purposes, as far as testing with or without Ini Priority.

A good test candidate game to confirm and compare before and after results is Dr. Mario with framebuffer ON and Ini OFF/ON.  You will get wild glitches with Ini Off.  And, you will yield proper graphics with Ini On.

Case by Case Scenario, depending on the Platform, you may gain FPS with Ini OFF/ON.  Be warned.  Games may NOT boot correctly, if you disable INI.  So, be prepared to revert, if need be...to restore accessibility to said games.